refactor: rename storage encoding module
